feat: support readme-styled codesamples
