+++ cmd/generate/codesamples.go
@@ -3,13 +3,13 @@ package generate
 import (
 	"context"
 	"fmt"
+	"github.com/speakeasy-api/speakeasy/internal/codesamples"
 
 	charm_internal "github.com/speakeasy-api/speakeasy/internal/charm"
 	"github.com/speakeasy-api/speakeasy/internal/charm/styles"
 	"github.com/speakeasy-api/speakeasy/internal/log"
 	"github.com/speakeasy-api/speakeasy/internal/model"
 	"github.com/speakeasy-api/speakeasy/internal/model/flag"
-	"github.com/speakeasy-api/speakeasy/internal/usagegen"
 )
 
 type codeSamplesFlags struct {
@@ -19,6 +19,7 @@ type codeSamplesFlags struct {
 	Langs      []string `json:"langs"`
 	ConfigPath string   `json:"config-path"`
 	Out        string   `json:"out"`
+	Style      string   `json:"style"`
 }
 
 var codeSamplesCmd = &model.ExecutableCommand[codeSamplesFlags]{
@@ -50,11 +51,25 @@ var codeSamplesCmd = &model.ExecutableCommand[codeSamplesFlags]{
 			Name:        "out",
 			Description: "write directly to a file instead of stdout",
 		},
+		flag.EnumFlag{
+			Name:          "style",
+			Description:   "the codeSamples style to generate, usually based on where the code samples will be used",
+			DefaultValue:  "standard",
+			AllowedValues: []string{"standard", "readme"},
+		},
 	},
 }
 
 func runCodeSamples(ctx context.Context, flags codeSamplesFlags) error {
-	_, err := usagegen.GenerateCodeSamplesOverlay(ctx, flags.Schema, flags.Header, flags.Token, flags.ConfigPath, flags.Out, flags.Langs, false)
+	var style codesamples.CodeSamplesStyle
+	switch flags.Style {
+	case "standard":
+		style = codesamples.Default
+	case "readme":
+		style = codesamples.ReadMe
+	}
+
+	_, err := codesamples.GenerateOverlay(ctx, flags.Schema, flags.Header, flags.Token, flags.ConfigPath, flags.Out, flags.Langs, false, style)
 
 	if err == nil {
 		locationString := "Overlay file written to stdout"
