+++ cs.yaml
@@ -1,0 +1,103 @@
+overlay: 1.0.0
+info:
+  title: CodeSamples overlay for ../openapi-ding.yaml
+  version: 0.0.0
+extends: file:///Users/robertcrumbaugh/code/openapi-ding.yaml
+actions:
+  - target: $["paths"]["/authentication"]["post"]
+    update:
+      x-readme:
+        code-samples:
+          - name: autentication
+            language: typescript
+            code: |-
+              import { SDK } from "openapi";
+
+              const sdk = new SDK({
+                apiKey: "YOUR_API_KEY",
+              });
+
+              async function run() {
+                const result = await sdk.otp.createAutentication({
+                  customerUuid: "eae192ab-9e1e-4b21-b5b1-bfcb79a32fcc",
+                  phoneNumber: "+1234567890",
+                });
+
+                // Handle the result
+                console.log(result)
+              }
+
+              run();
+  - target: $["paths"]["/check"]["post"]
+    update:
+      x-readme:
+        code-samples:
+          - name: check
+            language: typescript
+            code: |-
+              import { SDK } from "openapi";
+
+              const sdk = new SDK({
+                apiKey: "YOUR_API_KEY",
+              });
+
+              async function run() {
+                const result = await sdk.otp.check({
+                  customerUuid: "e0e7b0e9-739d-424b-922f-1c2cb48ab077",
+                  authenticationUuid: "8f1196d5-806e-4b71-9b24-5f96ec052808",
+                  checkCode: "123456",
+                });
+
+                // Handle the result
+                console.log(result)
+              }
+
+              run();
+  - target: $["paths"]["/retry"]["post"]
+    update:
+      x-readme:
+        code-samples:
+          - name: retry
+            language: typescript
+            code: |-
+              import { SDK } from "openapi";
+
+              const sdk = new SDK({
+                apiKey: "YOUR_API_KEY",
+              });
+
+              async function run() {
+                const result = await sdk.otp.retry({
+                  customerUuid: "a74ee547-564d-487a-91df-37fb25413a91",
+                  authenticationUuid: "3c8b3a46-881e-4cdd-93a6-f7f238bf020a",
+                });
+
+                // Handle the result
+                console.log(result)
+              }
+
+              run();
+  - target: $["paths"]["/lookup/{phone_number}"]["get"]
+    update:
+      x-readme:
+        code-samples:
+          - name: lookup
+            language: typescript
+            code: |-
+              import { SDK } from "openapi";
+
+              const sdk = new SDK({
+                apiKey: "YOUR_API_KEY",
+              });
+
+              async function run() {
+                const result = await sdk.lookup.lookup({
+                  customerUuid: "6e93aa15-9177-4d09-8395-b69ce50db1c8",
+                  phoneNumber: "<value>",
+                });
+
+                // Handle the result
+                console.log(result)
+              }
+
+              run();
